pyproject.toml
[tool.poetry]
name = "volttron-listener-agent"
version = "0.1.2"
description = "VOLTTRON listener agent.  Listenes to all published messages and writes to the log file."
authors = ["Craig Allwardt <craig.allwardt@pnnl.gov>"]
license = "Apache-2.0"
packages = [
    { include = "listener" }
]
homepage = "https://github.com/VOLTTRON/volttron-listener-agent"
repository = "https://github.com/VOLTTRON/volttron-listener-agent"

keywords = ["volttron", "agent", "messagebus", "application"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
requests = "^2.26.0"
#volttron-client = "^0.3.8"

[tool.poetry.group.dev.dependencies]
volttron-testing = {path = "../volttron-testing", develop = true}

[tool.poetry.dev-dependencies]
# formatting, quality, tests
pre-commit = "^2.17.0"
yapf = "^0.32.0"
toml = "^0.10.2"
isort = "^5.10.1"
safety = "^1.10.3"
mypy = "^0.942"
coverage = "^6.3.2"
volttron = {path="../volttron-core", develop=true}

[tool.poetry.scripts]
listener="listener.agent:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/'''
